+++ lib/src/foundation.dart
@@ -1,4 +1,5 @@
 const foundationSource = '''
 library foundation;
 export 'package:flutter/src/foundation/change_notifier.dart';
+export 'package:flutter/src/foundation/key.dart';
 ''';