+++ lib/src/widgets/framework.dart
@@ -7,10 +7,10 @@ import 'package:flutter_eval/src/foundation/diagnostics.dart';
 import 'package:flutter_eval/src/foundation/key.dart';
 
 const widgetsFrameworkSource = ''' 
-export 'package:flutter/foundation.dart' show FlutterError, ErrorSummary, ErrorDescription, ErrorHint, debugPrint, debugPrintStack;
-export 'package:flutter/foundation.dart' show VoidCallback, ValueChanged, ValueGetter, ValueSetter;
-export 'package:flutter/foundation.dart' show DiagnosticsNode, DiagnosticLevel;
-export 'package:flutter/foundation.dart' show Key, LocalKey, ValueKey;
+export 'package:flutter/foundation.dart' show FlutterError, ErrorSummary, ErrorDescription, ErrorHint,
+                    debugPrint, debugPrintStack, VoidCallback, ValueChanged, 
+                    ValueGetter, ValueSetter, DiagnosticsNode, DiagnosticLevel, 
+                    Key, LocalKey, ValueKey;
 ''';
 
 class $Widget implements Widget, $Instance {
@@ -26,7 +26,7 @@ class $Widget implements Widget, $Instance {
 
   @override
   $Value? $getProperty(Runtime runtime, String identifier) {
-    throw UnimplementedError();
+    throw UnimplementedError('Unknown property "$identifier"');
   }
 
   @override
@@ -360,7 +360,9 @@ class $State$bridge<T extends StatefulWidget> extends State<T> with $Bridge<Stat
             returns: BridgeTypeAnnotation($Widget.$type),
             params: [BridgeParameter('context', BridgeTypeAnnotation($BuildContext.$type), false)])),
       },
-      getters: {},
+      getters: {
+        'widget': BridgeMethodDef(BridgeFunctionDef(returns: BridgeTypeAnnotation(BridgeTypeRef.ref('T')))),
+      },
       setters: {},
       fields: {},
       bridge: true);
@@ -384,9 +386,14 @@ class $State$bridge<T extends StatefulWidget> extends State<T> with $Bridge<Stat
           super.initState();
           return null;
         });
+      case 'widget':
+        if (widget is $Instance) {
+          return widget as $Instance;
+        }
+        return $Widget.wrap(widget);
     }
 
-    throw UnimplementedError();
+    throw UnimplementedError('Unknown property "$identifier"');
   }
 
   @override
@@ -493,7 +500,7 @@ class $BuildContext implements BuildContext, $Instance {
   Widget get widget => $value.widget;
 }
 
-class $StatelessElement implements StatelessElement, $Instance {
+class $StatelessElement implements $Instance {
   const $StatelessElement.wrap(this.$value);
 
   @override
@@ -514,218 +521,6 @@ class $StatelessElement implements StatelessElement, $Instance {
   void $setProperty(Runtime runtime, String identifier, $Value value) {
     throw UnimplementedError();
   }
-
-  @override
-  void activate() => $value.activate();
-
-  @override
-  void attachNotificationTree() => $value.attachNotificationTree();
-
-  @override
-  void attachRenderObject(Object? newSlot) => $value.attachRenderObject(newSlot);
-
-  @override
-  Widget build() => $value.build();
-
-  @override
-  void deactivate() => $value.deactivate();
-
-  @override
-  void deactivateChild(Element child) => $value.deactivateChild(child);
-
-  @override
-  void debugDeactivated() => $value.debugDeactivated();
-
-  @override
-  List<DiagnosticsNode> debugDescribeChildren() => $value.debugDescribeChildren();
-
-  @override
-  bool get debugDoingBuild => $value.debugDoingBuild;
-
-  @override
-  void debugFillProperties(DiagnosticPropertiesBuilder properties) => $value.debugFillProperties(properties);
-
-  @override
-  String debugGetCreatorChain(int limit) => $value.debugGetCreatorChain(limit);
-
-  @override
-  List<Element> debugGetDiagnosticChain() => $value.debugGetDiagnosticChain();
-
-  @override
-  bool get debugIsDefunct => $value.debugIsDefunct;
-
-  @override
-  void debugVisitOnstageChildren(ElementVisitor visitor) => $value.debugVisitOnstageChildren(visitor);
-
-  @override
-  InheritedWidget dependOnInheritedElement(InheritedElement ancestor, {Object? aspect}) =>
-      $value.dependOnInheritedElement(ancestor, aspect: aspect);
-
-  @override
-  T? dependOnInheritedWidgetOfExactType<T extends InheritedWidget>({Object? aspect}) =>
-      $value.dependOnInheritedWidgetOfExactType<T>(aspect: aspect);
-
-  @override
-  int get depth => $value.depth;
-
-  @override
-  DiagnosticsNode describeElement(String name, {DiagnosticsTreeStyle style = DiagnosticsTreeStyle.errorProperty}) =>
-      $value.describeElement(name, style: style);
-
-  @override
-  List<DiagnosticsNode> describeMissingAncestor({required Type expectedAncestorType}) =>
-      $value.describeMissingAncestor(expectedAncestorType: expectedAncestorType);
-
-  @override
-  DiagnosticsNode describeOwnershipChain(String name) => $value.describeOwnershipChain(name);
-
-  @override
-  DiagnosticsNode describeWidget(String name, {DiagnosticsTreeStyle style = DiagnosticsTreeStyle.errorProperty}) =>
-      $value.describeWidget(name, style: style);
-
-  @override
-  void detachRenderObject() => $value.detachRenderObject();
-
-  @override
-  void didChangeDependencies() => $value.didChangeDependencies();
-
-  @override
-  bool get dirty => $value.dirty;
-
-  @override
-  void dispatchNotification(Notification notification) => $value.dispatchNotification(notification);
-
-  @override
-  T? findAncestorRenderObjectOfType<T extends RenderObject>() {
-    // TODO: implement findAncestorRenderObjectOfType
-    throw UnimplementedError();
-  }
-
-  @override
-  T? findAncestorStateOfType<T extends State<StatefulWidget>>() {
-    // TODO: implement findAncestorStateOfType
-    throw UnimplementedError();
-  }
-
-  @override
-  T? findAncestorWidgetOfExactType<T extends Widget>() {
-    // TODO: implement findAncestorWidgetOfExactType
-    throw UnimplementedError();
-  }
-
-  @override
-  RenderObject? findRenderObject() => $value.findRenderObject();
-
-  @override
-  T? findRootAncestorStateOfType<T extends State<StatefulWidget>>() {
-    // TODO: implement findRootAncestorStateOfType
-    throw UnimplementedError();
-  }
-
-  @override
-  void forgetChild(Element child) {
-    // TODO: implement forgetChild
-  }
-
-  @override
-  InheritedElement? getElementForInheritedWidgetOfExactType<T extends InheritedWidget>() {
-    // TODO: implement getElementForInheritedWidgetOfExactType
-    throw UnimplementedError();
-  }
-
-  @override
-  Element inflateWidget(Widget newWidget, Object? newSlot) => $value.inflateWidget(newWidget, newSlot);
-
-  @override
-  void markNeedsBuild() => $value.markNeedsBuild();
-
-  @override
-  void mount(Element? parent, Object? newSlot) => $value.mount(parent, newSlot);
-
-  @override
-  // TODO: implement owner
-  BuildOwner? get owner => throw UnimplementedError();
-
-  @override
-  void performRebuild() => $value.performRebuild();
-
-  @override
-  void reassemble() => $value.reassemble();
-
-  @override
-  void rebuild() => $value.rebuild();
-
-  @override
-  RenderObject? get renderObject => $value.renderObject;
-
-  @override
-  Size? get size => $value.size;
-
-  @override
-  Object? get slot => $value.slot;
-
-  @override
-  DiagnosticsNode toDiagnosticsNode({String? name, DiagnosticsTreeStyle? style}) {
-    // TODO: implement toDiagnosticsNode
-    throw UnimplementedError();
-  }
-
-  @override
-  String toStringDeep(
-      {String prefixLineOne = '', String? prefixOtherLines, DiagnosticLevel minLevel = DiagnosticLevel.debug}) {
-    // TODO: implement toStringDeep
-    throw UnimplementedError();
-  }
-
-  @override
-  String toStringShallow({String joiner = ', ', DiagnosticLevel minLevel = DiagnosticLevel.debug}) {
-    // TODO: implement toStringShallow
-    throw UnimplementedError();
-  }
-
-  @override
-  String toStringShort() {
-    // TODO: implement toStringShort
-    throw UnimplementedError();
-  }
-
-  @override
-  String toString({DiagnosticLevel minLevel = DiagnosticLevel.info}) => $value.toString(minLevel: minLevel);
-
-  @override
-  void unmount() => $value.unmount();
-
-  @override
-  void update(covariant StatelessWidget newWidget) => $value.update(newWidget);
-
-  @override
-  Element? updateChild(Element? child, Widget? newWidget, Object? newSlot) {
-    // TODO: implement updateChild
-    throw UnimplementedError();
-  }
-
-  @override
-  void updateSlotForChild(Element child, Object? newSlot) {
-    // TODO: implement updateSlotForChild
-  }
-
-  @override
-  void visitAncestorElements(bool Function(Element element) visitor) {
-    // TODO: implement visitAncestorElements
-  }
-
-  @override
-  void visitChildElements(ElementVisitor visitor) {
-    // TODO: implement visitChildElements
-  }
-
-  @override
-  void visitChildren(ElementVisitor visitor) {
-    // TODO: implement visitChildren
-  }
-
-  @override
-  Widget get widget => $value.widget;
 }
 
 class $StatefulElement implements $Instance {